Feature/qot 1581 daily check in3 (#1719)
* fixed conflict

*  pod update

* Add Bespoke to daily brief

* Handle Be spoke cell data

* Add contentview coolor

* Swift lint

* Add placeholder images

* Set access modifers

* Fix as per review commetns

* Fix as per review comments

*  disregard this commit

* whatshotlatest cell created

* question cell created

* random thoughts

*  good to know cell created

*  message from tignum cell created

* message from tignum

* updated admin

* cleaned up

* tweeked message from tignum

*  no more magic numbers

*  departure info added

* added QDMDailybriefbucket

* whatshot

*  gesture recogniser added to whatshot cell only

* returning cell

* DailyBriefView controller into tableview

*  created explore

* new files

* created view

*  changes

*  added feast for eyes

*  created three cases

*  explore cell done

* cleaned up

* Work in progress - Daily Check In

* cells

* applied one change

*  changes applied

* Working in progress - Daily Check in

* Updated viewmodel for daily checkin

* creagted dummy data

* wip tableview for dailycheck in

* Updated table view ReadinessDataTableView

*  minor setup

* WIP - Daily check in

* tableview

* WIP - Daily check in

* SHPI Cell created

* reated other cells

* numbers added

* changes

* WIP - Daily Check In

*  changes

* updated pods

* updated tbv cell

* Added Image url for the dailycheckIn
WIP - Daily checkin feature

* Added DailyBriefPeakPerformance and DailyBriefPeakPerformance Cell UI.
Added DailyBriefForwardEdge Cell.
Added DailyCheckInInsights2TBV cell.

* rating tbv sentences

*  podfile.lock

* looking for question ids

*  save

* Updated qot_dal in the repo.

* no database working

* changed

* corrected dateds in MYQOT

* nopreparations view repared in MYQOT

* no selection of cells in dailybrief

* WIP Dailybrief  UI.

* changed feast for eyes model

*  trying to update db

*  pod update

*  db built

* WIP DailyBrief Implementation.

*  updated tbv chekcin

* Added BeSpoke Bucket Implementation.
Modified BeSpoke Cell.

*  shpi in progress

* Removed unnecessary gestures.
Add Tignum message cell.

*  updates to impact readiness cell

* updated readiness cell

* merged explore

*  shpi cell complete

* impat readiness intro phrase

* WIP Daily Brief Implementation.

* subtitles and titles in order on impact readiness

* creating customized target view

* Impact readiness - Null state

* customize target labels

*  customize target almost done

* Updated About Me and Me At My Best buckets.

*  customize tafget failing to access button

*  finished peak performance daily checkin2

* TODO update the viewModel generation.

*  updated shpi indexing of response

* Leaders wisdom changes.

* updates

* wip differencekit implementation for all the buckets.

*  savec value

* working version of level 5

*  rounded save button

* Added view model for daily insights 2.

* updated level 5

* printing

* customized target localized strings

* saving value

* wip converting explore bucket to view model.
converted GetToLevel 5 and Leaders Wisdom to view model.
Modified whatshot bucket to view model.

* saving value of hours

* finalized

* cleared a bit

*  updated

* WIP DailyCheckIn 2 to viewmodel

* leader wisdom playing audio and video

*  cleaned

* from coach

*  level 5 ui

*  UI settings

* adjusted cell sizes

*  fixed explore

* MyPeakPerformance and DailyCheckin wip

* merged

*  going to merge

* level 5

*  saved screen

* solve UI finished

* sprint cell ui completed missing content collections

* work in progress

*  cleaned up

*  fixed sprint

*  getting titles from cms

* removed warnings.

*  adjusting layout

*  set selection to none

*  fixing name in leader wisdom

* updated qot_dal

* changes applied

* new changes

*  changes

*  level5 outlets collections

* updated qotdal

*  uödated solve bucket with titles

* fixed solve

* fixed all models

*  UI adjusted

* Code formatting, removed force wrap method.

* Updated review comments.

*  daily checkin1

* check it button working in solve

* daily checkin link and solve bucket complete

* update tbv possible

* My Peak Performance data Integration.

*  updated impact readiness

* changes applied

* customize target

* update

*  customize target

*  bon calcul

*  expanding bucket

* UI finished

*  cleaned up

*  cleaned up

* bad request

* removedobservr

* development team

+++ QOT/Sources/ImpactReadinessCellViewModel.swift
@@ -18,7 +18,7 @@ final class ImpactReadinessCellViewModel: BaseDailyBriefViewModel {
     var title: String?
     var dailyCheckImageView: URL?
     var readinessScore: Int?
-    var targetReferenceArray: [Int]?
+    var targetReferenceArray: [Double]?
     var impactDataModels: [ImpactDataViewModel]?
 
     struct ImpactDataViewModel {
@@ -29,7 +29,7 @@ final class ImpactReadinessCellViewModel: BaseDailyBriefViewModel {
     }
 
     // MARK: - Init
-    internal init(title: String?, dailyCheckImageView: URL?, howYouFeelToday: String?, asteriskText: String?, readinessScore: Int?, targetReferenceArray: [Int]?, impactDataModels: [ImpactDataViewModel]?, readinessIntro: String?, domainModel: QDMDailyBriefBucket?) {
+    internal init(title: String?, dailyCheckImageView: URL?, howYouFeelToday: String?, asteriskText: String?, readinessScore: Int?, targetReferenceArray: [Double]?, impactDataModels: [ImpactDataViewModel]?, readinessIntro: String?, domainModel: QDMDailyBriefBucket?) {
         self.targetReferenceArray = targetReferenceArray
         self.title = title
         self.dailyCheckImageView = dailyCheckImageView
